add "dirty mode" to build-backend
this is useful for when you want to add changes to your built version, for code that is using it through "npm link", without blowing away node_modules

+++ scripts/build-backend.sh
@@ -2,10 +2,16 @@
 
 set -e
 
-rm -rf dist
+if [ "$1" = "dirty" ]; then
+  echo "Dirty mode: not removing previous build files."
+else
+  rm -rf dist
+fi
+
 babel contracts --out-dir ./dist/contracts
 babel contracts/addresses --out-dir ./dist/contracts/addresses
 babel src --out-dir ./dist/src
+
 copyfiles \
   npm-readme.md \
   LICENSE \
@@ -17,5 +23,6 @@ copyfiles \
   contracts/addresses/daiV1.json \
   contracts/addresses/exchanges.json \
   dist
+
 mv dist/npm-readme.md dist/README.md
 cp package-lock.json dist/npm-shrinkwrap.json
